Added cross platform build support for SIMD instrinsics; depth camera output buffer dtype now matches d345i

+++ iceoryx_interfaces/src/cpp/include/camera_data/DepthFrameData_.hpp
@@ -14,6 +14,6 @@ namespace iceoryx_interfaces::camera {
         float depth_min;
         float depth_max;
 
-        float data[kFrameBufferElementsDepth];
+        uint16_t data[kFrameBufferElementsDepth];
     };
 }
